NetworkEdge: field-level fidelity fixes vs network-edgev1 spec
ASN localAsn/remoteAsn -> Long (BGP+VPN, fixes 4-byte ASN overflow); device create agreeOrderTerms + accountNumber String + aclDetails/projectId/tier/etc.; DeviceLinks metroLinks rename+reshape; getAllowedInterfaces; reload/upgrade history, interface stats, downtime notifications; backup restore path fix; vendorConfig fields; audit *DateTime mapping.

+++ equinix-sdk-java/src/main/java/api/equinix/javasdk/networkedge/model/BGPPeering.java
@@ -74,16 +74,16 @@ public interface BGPPeering {
     /**
      * <p>getLocalAsn.</p>
      *
-     * @return a {@link java.lang.Integer} object.
+     * @return a {@link java.lang.Long} object.
      */
-    Integer getLocalAsn();
+    Long getLocalAsn();
 
     /**
      * <p>getRemoteAsn.</p>
      *
-     * @return a {@link java.lang.Integer} object.
+     * @return a {@link java.lang.Long} object.
      */
-    Integer getRemoteAsn();
+    Long getRemoteAsn();
 
     /**
      * <p>getAuthenticationKey.</p>
