+++ equinix-sdk-java/src/main/java/api/equinix/javasdk/networkedge/model/VPN.java
@@ -79,9 +79,9 @@ public interface VPN {
     /**
      * <p>getRemoteAsn.</p>
      *
-     * @return a {@link java.lang.Integer} object.
+     * @return a {@link java.lang.Long} object.
      */
-    Integer getRemoteAsn();
+    Long getRemoteAsn();
 
     /**
      * <p>getRemoteIpAddress.</p>
@@ -100,9 +100,9 @@ public interface VPN {
     /**
      * <p>getLocalAsn.</p>
      *
-     * @return a {@link java.lang.Integer} object.
+     * @return a {@link java.lang.Long} object.
      */
-    Integer getLocalAsn();
+    Long getLocalAsn();
 
     /**
      * <p>getProjectId.</p>
