+++ equinix-sdk-java/src/main/java/api/equinix/javasdk/networkedge/model/implementation/LinkDevice.java
@@ -61,7 +61,7 @@ public class LinkDevice {
     private NetworkScope networkScope;
 
     @JsonProperty("asn")
-    private Integer asn;
+    private Long asn;
 
     @JsonProperty("ipAssigned")
     private String ipAssigned;
