+++ equinix-sdk-java/src/main/java/api/equinix/javasdk/networkedge/model/json/creators/BGPPeeringUpdaterJson.java
@@ -38,10 +38,10 @@ public class BGPPeeringUpdaterJson {
     private String remoteIpAddress;
 
     @JsonProperty("localAsn")
-    private Integer localAsn;
+    private Long localAsn;
 
     @JsonProperty("remoteAsn")
-    private Integer remoteAsn;
+    private Long remoteAsn;
 
     @JsonProperty("authenticationKey")
     private String authenticationKey;
