+++ equinix-sdk-java/src/main/java/api/equinix/javasdk/networkedge/model/json/creators/VPNUpdaterJson.java
@@ -47,7 +47,7 @@ public class VPNUpdaterJson {
     private String peerSharedKey;
 
     @JsonProperty("remoteAsn")
-    private Integer remoteAsn;
+    private Long remoteAsn;
 
     @JsonProperty("remoteIpAddress")
     private String remoteIpAddress;
@@ -56,7 +56,7 @@ public class VPNUpdaterJson {
     private String password;
 
     @JsonProperty("localAsn")
-    private Integer localAsn;
+    private Long localAsn;
 
     @JsonProperty("tunnelIp")
     private String tunnelIp;
